added display & Hittable changes
added display in SniperGame class. Now we can display Levels, but there
is still a lot actual code to do.

Also changed Hittable from a subclass of Actor to a subclass of Group,
as I suddenly realized an Actor is an ImageView object, which cannot
account for the hitbox easily.